sylvester: Initial import
The sylvester service derives its name from Sylvester, the mathematician
who brought the concept of simultaneous equations and matricies from
the east to the west.

The purpose of this service is to allow users to "enter the matrix",
i.e, to view and edit the source code for the applications that
make up their system, easily.

Call DownloadSourcesForPid to look up the source code for a particular
process-id and use apt-get source to download its sources and open
a gnome-builder window.

DownloadSourcesForSelectedWindow is a debugging function to select
arbitrary windows and get their process-id.

The service will emit the "RotateIn" signal when two windows are to
be rotated as a transition.