service: Allow lessons to be loaded from non-resource sources
The service will first check if it was invoked with
--lessons-file. Because option parsing isn't really supported
for GApplications, this is effectively an undocumented option.

The service will then check if there is a ~/.data/lessons.json
in place.

If either one of those files is present, it will be loaded and
validated. If the file is present but has some invalid sections,
those sections will be removed and warnings will be added
to the descriptor. These can be queried with GetWarnings by
a client side. For instance, those prototyping by
using the chatbox can quickly check if there was any invalid
entries just by opening the chatbox as opposed to
having to dig through system logs.

Obviously, the preferable case is for a file to be loaded from
the resource and this is what will happen on the default image. But
for now, it is better for external developers to be able to
quickly iterate on their own files.