service: Add RegisterClue method
RegisterClue takes a tuple of arity two  containing both the
data type of the clue and the content of the clue. This can be
read by other applications to display certain custom content.

Currently we support text clues and also clues which are image
paths, given that the user may wish to store screenshots of
some sort.