service: Rework the way that lesson attempts happen
Instead of having fixed steps of text input, have a pipeline which
is specified by an initial input step (which can be custom specified
according to the json file). Then have a series of "mappers" which
transform the input into some discrete output, each taking the
return value of the last as its own input.

For instance, you might have a mapper which takes the input and
uses it to execute a command, or another mapper which takes
some text and matches it against some regex. Or another mapper
which just ignores its input and executes an external command
to check a side effect. These mappers can be chained in any
length.

Once the pipeline is complete, the result is checked against the
keys in "effects" to see what to do next. The must be a key in
"effects" for every possible output in the output space
of the last pipeline entry.