service: Add the concept of events that are external to the service
The service might end up listening for external events if a lesson
plan requires it. If that is the case, we need to maintain some state
on the service side temporarily to see when particular events have
happened. If a client makes a request to the service again on the
same lesson by "attempting" it with an empty message, we return a
response based on the events that have been satisfied up to this
point