service: Add support for lessons on the python runtime
Here, we add a new type of shell to supervise, being the Python REPL.
We cannot just use the Python REPL directly because of Python's
output buffering. It turns out that Python uses the standard error to
print newlines on the REPL so trying to do a non-blocking read directly
won't work. Instead, we force unbuffered mode and then spawn a new
InteractiveConsole instance.

We also want to be able to access GApplication methods which requires
a GApplication and GMainLoop instance to be running. We do this by
pre-seeding the interpreter with some code that starts a new GApplication
and runs the main loop