bin: Set HOME explicitly
In the case that we were run through D-Bus activation, we won't
be run through a shell, in which case this won't be set for us.
Certain lessons rely on this to be set, so we can just use
GLib.get_home_dir() to find out its value. That function will
read /etc/passwd if HOME isn't set.

https://phabricator.endlessm.com/T14922