controller: Add until_match_stdout and until_match_stderr conditions
When running interactive shells on slow systems, it is possible that
it could take a little longer than expected for all the output
to be printed. Add an until_match_stdout and until_match_stderr
condition which causes the matching process to consume more output
and repeat until a regex is satisfied up to repeat_times times.